app/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
DateTime==5.5
idna==3.7
pyaes==1.6.1
pyasn1==0.6.0
pytz==2024.1
requests==2.32.3
rsa==4.9
setuptools==71.1.0
Telethon==1.36.0
urllib3==2.2.2
zope.interface==6.4.post2
python-dotenv
